+++ db/migrate/20150505110742_create_providers.rb
@@ -1,0 +1,8 @@
+class CreateProviders < ActiveRecord::Migration
+  def change
+    create_table :providers do |t|
+      t.string :name, limit: 20, null: false, unique: true
+      t.timestamps
+    end
+  end
+end