+++ spec/acceptance/users_spec.rb
@@ -2,7 +2,7 @@ require 'documentation_helper'
 
 resource "Users" do
   let(:json) { JSON.parse(response_body) }
-  let(:current_user) { create(:user) }
+  let(:current_user) { create(:user, :with_email, :with_password) }
   let(:auth_header) { "Token token=#{current_user.auth_token}" }
 
   get "/v1/users/:id" do
@@ -14,6 +14,30 @@ resource "Users" do
     response_field :id, "ID of the user"
     response_field :username, "Username of the user"
     response_field :image, "URL of the user image"
+    response_field :email, "Email of user (max. 255 characters)"
+    response_field :phone, "Phone number of user (max. 35 characters)"
+    response_field :locationId, "Home location of user"
+    response_field :publicUrl, "Public URL for user on Timeslot (max. 255 chars)"
+    response_field :push, "Send push Notifications (true/false)"
+    response_field :slotDefaultDuration, "Default Slot Duration in seconds"
+    response_field :slotDefaultTypeId, "Default Slot Type - WIP"
+    response_field :slotDefaultLocationId, "Default Slot Location ID - WIP"
+    response_field :defaultPrivateAlerts,
+                   "Default alerts for private slots of this user"
+    response_field :defaultOwnFriendslotAlerts,
+                   "Default alerts for the friendslots of this user"
+    response_field :defaultOwnPublicAlerts,
+                   "Default alerts for the public slots of this user"
+    response_field :defaultFriendsFriendslotAlerts,
+                   "Default alerts for the friendslots from friends of this user"
+    response_field :defaultFriendsPublicAlerts,
+                   "Default alerts for the public slots from friends of this user"
+    response_field :defaultReslotAlerts,
+                   "Default alerts for the reslots of this user"
+    response_field :defaultGroupAlerts,
+                   "Default alerts for all groupslots of this user" \
+                   " where no specific alert is set. Groupslots" \
+                   " may also have their own default alerts per group"
     response_field :createdAt, "Creation of user"
     response_field :updatedAt, "Latest update of user in db"
     response_field :deletedAt, "Deletion of user"
@@ -40,8 +64,8 @@ resource "Users" do
 
     parameter :username, "Username of user (max. 50 characters)",
               required: true
-    parameter :email, "Email of user (max. 254 characters)",
-              required: true
+    parameter :email, "Email of user (max. 254 characters)"
+    parameter :phone, "Phone number of user (max. 35 characters)"
     parameter :password, "Password for user (min. 5 & max. 72 characters)",
               required: true
 
@@ -52,7 +76,8 @@ resource "Users" do
     let(:password) { "secret-thing" }
 
     example "User signup / Create user", document: :v1 do
-      explanation "returns 422 if parameters are missing\n\n" \
+      explanation "Either an email or phone number must be provided\n\n" \
+                  "returns 422 if parameters are missing\n\n" \
                   "returns 422 if parameters are invalid"
       do_request
 
@@ -61,7 +86,6 @@ resource "Users" do
       expect(json).to have_key 'username'
       expect(json).to have_key 'email'
       expect(json).to have_key 'authToken'
-
     end
   end
 
@@ -75,7 +99,7 @@ resource "Users" do
     response_field :authToken, "Authentication Token for the user to be set" \
                                " as a HTTP header in subsequent requests"
 
-    let(:user) { create(:user, password: "timeslot") }
+    let(:user) { create(:user, :with_email, password: "timeslot") }
     let(:email) { user.email }
     let(:password) { "timeslot" }
 
@@ -114,7 +138,7 @@ resource "Users" do
     parameter :email, "Email of the user for whom to reset password",
               required: true
 
-    let(:user) { create(:user, password: "nottimeslot") }
+    let(:user) { create(:user, :with_email, password: "nottimeslot") }
     let(:email) { user.email }
 
     example "Reset password", document: :v1 do
@@ -141,6 +165,15 @@ resource "Users" do
     describe "Update current users data" do
 
       parameter :username, "Updated username of user (max. 50 characters)"
+      parameter :email, "Email of user (max. 255 characters)"
+      parameter :phone, "Phone number of user (max. 35 characters)"
+      parameter :locationId, "Home location of user"
+      parameter :image, "URL of the user image"
+      parameter :publicUrl, "Public URL for user on Timeslot (max. 255 chars)"
+      parameter :push, "Send push Notifications (true/false)"
+      parameter :slotDefaultDuration, "Default Slot Duration in seconds"
+      parameter :slotDefaultTypeId, "Default Slot Type - WIP"
+      parameter :slotDefaultLocationId, "Default Slot Location ID - WIP"
       parameter :defaultPrivateAlerts,
                 "Default alerts for private slots of this user"
       parameter :defaultOwnFriendslotAlerts,